backend/requirements.txt
# FastAPI e servidor
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Autenticação e segurança
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0

# Banco de dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Validação e utilitários
email-validator>=2.0.0

# HTTP requests
requests>=2.31.0
httpx>=0.24.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
